Get Packages by biocViews

get_packages_by_view() returns package names associated with a single biocViews term.

get_packages_by_views() returns packages names at the intersection of multiple biocViews terms.

Usage

get_packages_by_view(view, pkg_list = NULL)

get_packages_by_views(views, pkg_list = NULL)

Arguments

view

A single biocView term.

pkg_list

Value of a call to biocPkgList(). If NULL (default), will call biocPkgList() internally. See Details.

views

A character vector of biocView terms.

Value

Character vector of package names.

Details

Calling BiocPkgTools::biocPkgList() and passing the result to get_packages_by_view() or get_packages_by_views() is more efficient if you are making multiple calls. See vignette 'Optimisations' for a more comprehensive discussion and demonstration.

Examples

get_packages_by_view("Spatial")
#> 'getOption("repos")' replaces Bioconductor standard repositories, see
#> 'help("repositories", package = "BiocManager")' for details.
#> Replacement repositories:
#>     CRAN: https://p3m.dev/cran/__linux__/noble/latest
#>  [1] "alabaster.sfe"            "Banksy"                  
#>  [3] "BatchSVG"                 "betaHMM"                 
#>  [5] "BulkSignalR"              "CARDspa"                 
#>  [7] "clustSIGNAL"              "concordexR"              
#>  [9] "CTSV"                     "cytoviewer"              
#> [11] "DESpace"                  "escheR"                  
#> [13] "FuseSOM"                  "GeomxTools"              
#> [15] "ggsc"                     "ggspavis"                
#> [17] "HiCPotts"                 "hoodscanR"               
#> [19] "HuBMAPR"                  "imcRtools"               
#> [21] "jazzPanda"                "knowYourCG"              
#> [23] "lisaClust"                "miRspongeR"              
#> [25] "mistyR"                   "mitology"                
#> [27] "MoleculeExperiment"       "nnSVG"                   
#> [29] "OSTA.data"                "pengls"                  
#> [31] "poem"                     "RegionalST"              
#> [33] "retrofit"                 "scatterHatch"            
#> [35] "sccomp"                   "scDesign3"               
#> [37] "scider"                   "SEraster"                
#> [39] "shinyDSP"                 "signifinder"             
#> [41] "simpleSeg"                "smoothclust"             
#> [43] "smoppix"                  "sosta"                   
#> [45] "SpaceTrooper"             "spacexr"                 
#> [47] "SpaNorm"                  "spARI"                   
#> [49] "spaSim"                   "SpatialDecon"            
#> [51] "SpatialExperiment"        "SpatialExperimentIO"     
#> [53] "spatialFDA"               "SpatialFeatureExperiment"
#> [55] "SpatialOmicsOverlay"      "spatialSimGP"            
#> [57] "SPIAT"                    "spicyR"                  
#> [59] "spoon"                    "SpotClean"               
#> [61] "SPOTlight"                "SpotSweeper"             
#> [63] "standR"                   "Statial"                 
#> [65] "stJoincount"              "stPipe"                  
#> [67] "SVP"                      "tidySpatialExperiment"   
#> [69] "tomoda"                   "tomoseqr"                
#> [71] "tpSVG"                    "VisiumIO"                
#> [73] "visiumStitched"           "Voyager"                 
#> [75] "XeniumIO"                 "CatsCradle"              
#> [77] "scFeatures"               "SpaceMarkers"            
#> [79] "spatialHeatmap"          
get_packages_by_views(c("Spatial", "SingleCell"))
#> 'getOption("repos")' replaces Bioconductor standard repositories, see
#> 'help("repositories", package = "BiocManager")' for details.
#> Replacement repositories:
#>     CRAN: https://p3m.dev/cran/__linux__/noble/latest
#>  [1] "Banksy"                "CARDspa"               "clustSIGNAL"          
#>  [4] "concordexR"            "cytoviewer"            "DESpace"              
#>  [7] "escheR"                "FuseSOM"               "ggsc"                 
#> [10] "ggspavis"              "hoodscanR"             "HuBMAPR"              
#> [13] "imcRtools"             "knowYourCG"            "lisaClust"            
#> [16] "miRspongeR"            "mistyR"                "mitology"             
#> [19] "nnSVG"                 "OSTA.data"             "poem"                 
#> [22] "retrofit"              "scatterHatch"          "sccomp"               
#> [25] "scDesign3"             "SEraster"              "signifinder"          
#> [28] "simpleSeg"             "smoothclust"           "smoppix"              
#> [31] "SpaceTrooper"          "spacexr"               "SpatialExperiment"    
#> [34] "SpatialExperimentIO"   "SPIAT"                 "spicyR"               
#> [37] "spoon"                 "SpotClean"             "SPOTlight"            
#> [40] "Statial"               "stPipe"                "SVP"                  
#> [43] "tidySpatialExperiment" "VisiumIO"              "XeniumIO"             
#> [46] "CatsCradle"            "scFeatures"            "SpaceMarkers"         
#> [49] "spatialHeatmap"