Get Packages by biocViews

get_packages_by_view() returns package names associated with a single biocViews term.

get_packages_by_views() returns packages names at the intersection of multiple biocViews terms.

Usage

get_packages_by_view(view, pkg_list = NULL)

get_packages_by_views(views, pkg_list = NULL)

Arguments

view

A single biocView term.

pkg_list

Value of a call to biocPkgList(). If NULL (default), will call biocPkgList() internally. See Details.

views

A character vector of biocView terms.

Value

Character vector of package names.

Details

Calling BiocPkgTools::biocPkgList() and passing the result to get_packages_by_view() or get_packages_by_views() is more efficient if you are making multiple calls. See vignette 'Optimisations' for a more comprehensive discussion and demonstration.

Examples

get_packages_by_view("Spatial")
#> 'getOption("repos")' replaces Bioconductor standard repositories, see
#> 'help("repositories", package = "BiocManager")' for details.
#> Replacement repositories:
#>     CRAN: https://p3m.dev/cran/__linux__/noble/latest
#>  [1] "alabaster.sfe"            "Banksy"                  
#>  [3] "BatchSVG"                 "betaHMM"                 
#>  [5] "BulkSignalR"              "CARDspa"                 
#>  [7] "CatsCradle"               "clustSIGNAL"             
#>  [9] "concordexR"               "CTSV"                    
#> [11] "cytoviewer"               "DESpace"                 
#> [13] "escheR"                   "FuseSOM"                 
#> [15] "GeomxTools"               "ggsc"                    
#> [17] "ggspavis"                 "HiCPotts"                
#> [19] "hoodscanR"                "HuBMAPR"                 
#> [21] "imcRtools"                "jazzPanda"               
#> [23] "knowYourCG"               "lisaClust"               
#> [25] "miRspongeR"               "mistyR"                  
#> [27] "mitology"                 "MoleculeExperiment"      
#> [29] "nnSVG"                    "OSTA.data"               
#> [31] "pengls"                   "poem"                    
#> [33] "RegionalST"               "retrofit"                
#> [35] "scatterHatch"             "sccomp"                  
#> [37] "scDesign3"                "scider"                  
#> [39] "SEraster"                 "shinyDSP"                
#> [41] "signifinder"              "simpleSeg"               
#> [43] "smoothclust"              "smoppix"                 
#> [45] "sosta"                    "SpaceMarkers"            
#> [47] "SpaceTrooper"             "spacexr"                 
#> [49] "SpaNorm"                  "spARI"                   
#> [51] "spaSim"                   "SpatialDecon"            
#> [53] "SpatialExperiment"        "SpatialExperimentIO"     
#> [55] "spatialFDA"               "SpatialFeatureExperiment"
#> [57] "SpatialOmicsOverlay"      "spatialSimGP"            
#> [59] "SPIAT"                    "spicyR"                  
#> [61] "spoon"                    "SpotClean"               
#> [63] "SPOTlight"                "SpotSweeper"             
#> [65] "standR"                   "Statial"                 
#> [67] "stJoincount"              "stPipe"                  
#> [69] "SVP"                      "tidySpatialExperiment"   
#> [71] "tomoda"                   "tomoseqr"                
#> [73] "tpSVG"                    "VisiumIO"                
#> [75] "visiumStitched"           "Voyager"                 
#> [77] "XeniumIO"                 "scFeatures"              
#> [79] "spatialHeatmap"          
get_packages_by_views(c("Spatial", "SingleCell"))
#> 'getOption("repos")' replaces Bioconductor standard repositories, see
#> 'help("repositories", package = "BiocManager")' for details.
#> Replacement repositories:
#>     CRAN: https://p3m.dev/cran/__linux__/noble/latest
#>  [1] "Banksy"                "CARDspa"               "CatsCradle"           
#>  [4] "clustSIGNAL"           "concordexR"            "cytoviewer"           
#>  [7] "DESpace"               "escheR"                "FuseSOM"              
#> [10] "ggsc"                  "ggspavis"              "hoodscanR"            
#> [13] "HuBMAPR"               "imcRtools"             "knowYourCG"           
#> [16] "lisaClust"             "miRspongeR"            "mistyR"               
#> [19] "mitology"              "nnSVG"                 "OSTA.data"            
#> [22] "poem"                  "retrofit"              "scatterHatch"         
#> [25] "sccomp"                "scDesign3"             "SEraster"             
#> [28] "signifinder"           "simpleSeg"             "smoothclust"          
#> [31] "smoppix"               "SpaceMarkers"          "SpaceTrooper"         
#> [34] "spacexr"               "SpatialExperiment"     "SpatialExperimentIO"  
#> [37] "SPIAT"                 "spicyR"                "spoon"                
#> [40] "SpotClean"             "SPOTlight"             "Statial"              
#> [43] "stPipe"                "SVP"                   "tidySpatialExperiment"
#> [46] "VisiumIO"              "XeniumIO"              "scFeatures"           
#> [49] "spatialHeatmap"